requirements.txt
flask
openai
faiss-cpu
python-dotenv
pymupdf
sentence-transformers